+++ playground/src/app/search/index.tsx
@@ -8,8 +8,8 @@ export const Search = () => {
 
   const search = useSearch<{ some: "thing" }>({
     _searchKey:
-      "k19b3p50ue6irgvnp-1.a1.typesense.net:YjY3bk8zVDFOY05mbTlCczk4U1dpU2Nqa2FTTHJKNXBxRHJPSWlHcTVRRT1WRnRteyJmaWx0ZXJfYnkiOiJfY29sbGVjdGlvbklkOnFmZDVva2FnaTA0eHc0ZmNmZjV6d3UyYSJ9:dbjFgtnqgk7aCqonrqA8Z__tm8owg2lnuah5n62mm56qqc7__zhzxbpr40ied562l74zvrdwg",
-    queryBy: ["_title", "content"],
+      "k19b3p50ue6irgvnp-1.a1.typesense.net:iH0LYufEw4sAk3aihviwX9tthXlzy6t8:RzwcwGKShcB0pBmOiUseY__6cilu6IceejaWHUYQoIWC__GOSCkL1oxXpFlktOiuZY_",
+    queryBy: ["_title", "body", "excerpt"],
     saveRecentSearches: {
       getStorage: () => window.localStorage,
       key: "recent-searches",
@@ -30,19 +30,28 @@ export const Search = () => {
         <SearchBox.Root search={search}>
           <SearchBox.Input autoFocus />
           <SearchBox.Placeholder asChild>
-            <SearchBox.HitsList asChild>
+            <SearchBox.HitList asChild>
               <ul>
                 <h3>Recent Searches</h3>
                 {search.recentSearches?.hits?.map((hit) => {
+                  let pathname = hit.document._slug;
+                  const bodyHighlight = hit._getFieldHighlight("body");
+                  if (
+                    bodyHighlight?.highlightedField?._type ===
+                      "rich-text-section" &&
+                    bodyHighlight.highlightedField._id
+                  ) {
+                    pathname += `#${bodyHighlight.highlightedField._id}`;
+                  }
                   return (
                     <li key={hit._key}>
                       <SearchBox.HitItem
                         hit={hit}
-                        href={`/doc/${hit.document._id}`}
+                        href={`/docs/${pathname}`}
                         className={s.hit}
                       >
                         <SearchBox.HitSnippet fieldPath="_title" />
-                        <SearchBox.HitSnippet fieldPath="content" />
+                        <SearchBox.HitSnippet fieldPath="body" />
                       </SearchBox.HitItem>
 
                       <button
@@ -56,30 +65,42 @@ export const Search = () => {
                   );
                 })}
               </ul>
-            </SearchBox.HitsList>
+            </SearchBox.HitList>
           </SearchBox.Placeholder>
           <SearchBox.Empty>
             <div>No results found for {search.query}</div>
           </SearchBox.Empty>
-          <SearchBox.HitsList asChild>
+          <SearchBox.HitList asChild>
             <ul style={{ maxHeight: 200, overflowY: "auto" }}>
               {search.result?.hits.map((hit) => {
+                let pathname = hit.document._slug;
+                const bodyHighlight = hit._getFieldHighlight("body");
+                if (
+                  bodyHighlight?.highlightedField?._type ===
+                    "rich-text-section" &&
+                  bodyHighlight.highlightedField._id
+                ) {
+                  pathname += `#${bodyHighlight.highlightedField._id}`;
+                }
                 return (
                   <li key={hit._key}>
                     <SearchBox.HitItem
                       key={hit._key}
                       hit={hit}
-                      href={`/doc/${hit.document._id}`}
+                      href={`/docs/${pathname}`}
                       className={s.hit}
                     >
                       <SearchBox.HitSnippet fieldPath="_title" />
-                      <SearchBox.HitSnippet fieldPath="content" />
+                      <SearchBox.HitSnippet
+                        fieldPath="body"
+                        fallbackFieldPaths={["excerpt"]}
+                      />
                     </SearchBox.HitItem>
                   </li>
                 );
               })}
             </ul>
-          </SearchBox.HitsList>
+          </SearchBox.HitList>
         </SearchBox.Root>
       )}
     </>